pyproject.toml
[tool.poetry]
name = "aws-cli-chatbot"
version = "0.1.0"
description = ""
authors = ["Remco Kersten"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.14"
langchain-core = "^0.2.11"
langchain-community = "^0.2.6"
langchainhub = "^0.1.20"
langchain-experimental = "^0.0.62"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
